musl: Show error when used with multilibs
As discussed in the bugzilla entry, musl doesn't work with multilibs.

I originally sent this patch in 2020 but was told that support was being
added. I recently revisited the bug and retested a basic config which still
fails the way it did in 2019. Since clearly nobody is using this, make it clear
we don't support it. If anyone does add support, we can remove this error very
easily.

[YOCTO #13122]

(From OE-Core rev: 8f47589af53fb9f731848bec512a21cee0ae29bf)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>